refactor proxy context
